Moved module tests into seperate files
